Выполнено частичное структурирование файлов - по совету наставника были отделены шаблоны от логики и разнесены по отдельным файлам в пределах одной директории
