+++ src/presenter/route-presenter.js
@@ -1,8 +1,8 @@
-import EditEventFormView from '../view/edit-event-form-view.js';
-import EventView from '../view/event-view.js';
-import EventsListItemView from '../view/events-list-item-view.js';
-import EventsListView from '../view/events-list-view.js';
-import SortFormView from '../view/sort-form-view.js';
+import EditEventFormView from '../view/edit_event_form/edit-event-form-view.js';
+import EventView from '../view/event/event-view.js';
+import EventsListItemView from '../view/events_list_item/events-list-item-view.js';
+import EventsListView from '../view/event_list/events-list-view.js';
+import SortFormView from '../view/sort_form/sort-form-view.js';
 import {render} from '../render.js';
 
 const EVENTS_AMOUNT = 3;
