+++ src/view/event_list/events-list-view.js
@@ -1,0 +1,20 @@
+import {createElement} from '../../render.js';
+import {createEventsListTemplate} from './events-list.tpl.js';
+
+export default class EventsListView {
+  getTemplate() {
+    return createEventsListTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
