+++ src/view/event_list/events-list.tpl.js
@@ -1,0 +1,5 @@
+const createEventsListTemplate = () => (
+  '<ul class="trip-events__list"></ul>'
+);
+
+export {createEventsListTemplate};
