+++ src/view/events_list_item/events-list-item.tpl.js
@@ -1,0 +1,5 @@
+const createEventsListItemTemplate = () => (
+  '<li class="trip-events__item"></li>'
+);
+
+export {createEventsListItemTemplate};
